requirements.txt
dash>=3.0.4
dash-bootstrap-components>=2.0.0
dash-leaflet>=1.0.13
dash-extensions>=1.0.0
plotly>=5.18.0
requests>=2.31.0
pandas>=2.1.4
python-dotenv>=1.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
